Finished the front end for rankings
